Move test files to test sources
